Move testing utils from core to testing
