+++ src/ophyd_async/testing/__init__.py
@@ -1,22 +1,33 @@
-import asyncio
+from ._assert import (
+    assert_configuration,
+    assert_emitted,
+    assert_reading,
+    assert_value,
+)
+from ._mock_signal_utils import (
+    callback_on_mock_put,
+    get_mock,
+    get_mock_put,
+    mock_puts_blocked,
+    reset_mock_put_calls,
+    set_mock_put_proceeds,
+    set_mock_value,
+    set_mock_values,
+)
+from ._wait_for_pending import wait_for_pending_wakeups
 
-
-async def wait_for_pending_wakeups(max_yields=20, raise_if_exceeded=True):
-    """Allow any ready asyncio tasks to be woken up.
-
-    Used in:
-
-    - Tests to allow tasks like ``set()`` to start so that signal
-      puts can be tested
-    - `observe_value` to allow it to be wrapped in `asyncio.wait_for`
-      with a timeout
-    """
-    loop = asyncio.get_event_loop()
-    # If anything has called loop.call_soon or is scheduled a wakeup
-    # then let it run
-    for _ in range(max_yields):
-        await asyncio.sleep(0)
-        if not loop._ready:  # type: ignore # noqa: SLF001
-            return
-    if raise_if_exceeded:
-        raise RuntimeError(f"Tasks still scheduling wakeups after {max_yields} yields")
+__all__ = [
+    "assert_configuration",
+    "assert_emitted",
+    "assert_reading",
+    "assert_value",
+    "callback_on_mock_put",
+    "get_mock",
+    "get_mock_put",
+    "mock_puts_blocked",
+    "reset_mock_put_calls",
+    "set_mock_put_proceeds",
+    "set_mock_value",
+    "set_mock_values",
+    "wait_for_pending_wakeups",
+]
