+++ tests/core/test_device_collector.py
@@ -9,9 +9,9 @@ from ophyd_async.core import (
     Device,
     DeviceCollector,
     NotConnected,
-    set_mock_value,
 )
 from ophyd_async.epics import motor
+from ophyd_async.testing import set_mock_value
 
 
 class FailingDevice(Device):
