+++ tests/core/test_signal.py
@@ -16,20 +16,22 @@ from ophyd_async.core import (
     SignalRW,
     SoftSignalBackend,
     StandardReadable,
-    assert_configuration,
-    assert_reading,
-    assert_value,
-    callback_on_mock_put,
     set_and_wait_for_other_value,
     set_and_wait_for_value,
-    set_mock_put_proceeds,
-    set_mock_value,
     soft_signal_r_and_setter,
     soft_signal_rw,
     wait_for_value,
 )
 from ophyd_async.core import StandardReadableFormat as Format
 from ophyd_async.epics.core import epics_signal_r, epics_signal_rw
+from ophyd_async.testing import (
+    assert_configuration,
+    assert_reading,
+    assert_value,
+    callback_on_mock_put,
+    set_mock_put_proceeds,
+    set_mock_value,
+)
 
 
 def num_occurrences(substring: str, string: str) -> int:
