+++ tests/epics/adcore/test_scans.py
@@ -17,9 +17,9 @@ from ophyd_async.core import (
     StandardDetector,
     StandardFlyer,
     TriggerInfo,
-    set_mock_value,
 )
 from ophyd_async.epics import adcore, adsimdetector
+from ophyd_async.testing import set_mock_value
 
 
 class DummyTriggerLogic(FlyerController[int]):
