+++ tests/epics/adpilatus/test_pilatus.py
@@ -7,9 +7,9 @@ import pytest
 from ophyd_async.core import (
     DetectorTrigger,
     TriggerInfo,
-    set_mock_value,
 )
 from ophyd_async.epics import adcore, adpilatus
+from ophyd_async.testing import set_mock_value
 
 
 @pytest.fixture
