+++ tests/epics/advimba/test_vimba.py
@@ -4,7 +4,6 @@ from ophyd_async.core import (
     DetectorTrigger,
     PathProvider,
     TriggerInfo,
-    set_mock_value,
 )
 from ophyd_async.epics import advimba
 from ophyd_async.epics.advimba import (
@@ -12,6 +11,7 @@ from ophyd_async.epics.advimba import (
     VimbaOnOff,
     VimbaTriggerSource,
 )
+from ophyd_async.testing import set_mock_value
 
 
 @pytest.fixture
