+++ tests/epics/eiger/test_eiger_controller.py
@@ -5,11 +5,9 @@ from pytest import fixture, raises
 from ophyd_async.core import (
     DeviceCollector,
     TriggerInfo,
-    callback_on_mock_put,
-    get_mock_put,
-    set_mock_value,
 )
 from ophyd_async.epics.eiger import EigerController, EigerDriverIO
+from ophyd_async.testing import callback_on_mock_put, get_mock_put, set_mock_value
 
 DriverAndController = tuple[EigerDriverIO, EigerController]
