+++ tests/epics/test_motor.py
@@ -9,16 +9,18 @@ from ophyd_async.core import (
     CALCULATE_TIMEOUT,
     AsyncStatus,
     DeviceCollector,
-    callback_on_mock_put,
-    get_mock_put,
-    mock_puts_blocked,
     observe_value,
-    set_mock_put_proceeds,
-    set_mock_value,
     soft_signal_rw,
 )
 from ophyd_async.epics import motor
-from ophyd_async.testing import wait_for_pending_wakeups
+from ophyd_async.testing import (
+    callback_on_mock_put,
+    get_mock_put,
+    mock_puts_blocked,
+    set_mock_put_proceeds,
+    set_mock_value,
+    wait_for_pending_wakeups,
+)
 
 
 @pytest.fixture
