+++ tests/sim/test_streaming_plan.py
@@ -3,9 +3,9 @@ from collections import defaultdict
 from bluesky import plans as bp
 from bluesky.run_engine import RunEngine
 
-from ophyd_async.core import assert_emitted
 from ophyd_async.plan_stubs import ensure_connected
 from ophyd_async.sim.demo import PatternDetector
+from ophyd_async.testing import assert_emitted
 
 
 # NOTE the async operations with h5py are non-trival
@@ -43,12 +43,13 @@ async def test_streaming_plan(RE: RunEngine, sim_pattern_detector: PatternDetect
 
 async def test_plan(RE: RunEngine, sim_pattern_detector: PatternDetector):
     docs = defaultdict(list)
+    RE.subscribe(lambda name, doc: docs[name].append(doc))
 
     def plan():
         yield from ensure_connected(sim_pattern_detector, mock=True)
         yield from bp.count([sim_pattern_detector])
 
-    RE(plan(), lambda name, doc: docs[name].append(doc))
+    RE(plan())
 
     assert_emitted(
         docs, start=1, descriptor=1, stream_resource=2, stream_datum=2, event=1, stop=1
